+++ .github/workflows/test-e2e-create.yml
@@ -40,11 +40,14 @@ jobs:
         with:
           repository: kyma-project/template-operator
           path: ./template-operator/
-      - name: export template-operator URL
+      - name: Copy template-operator to e2e folder
+        run: |
+          cp -r ./template-operator/* ./tests/e2e/create/
+      - name: Set template-operator github URL
         run: |
           cd ./template-operator
           echo "TEST_REPOSITORY_URL=$(git remote get-url origin)" >> "$GITHUB_ENV"
-      - name: Set up k3d
+      - name: Install k3d
         run: wget -qO - https://raw.githubusercontent.com/k3d-io/k3d/main/install.sh | TAG=$K3D_VERSION bash
       - name: Provision OCI Registry
         run: |
